Delete old releases (#21)
Releasing the wheels has been failing for a month now
https://github.com/llvm/torch-mlir-release/actions/runs/19032410905/job/54349329471#step:7:30
due to the GH limit that each release artifact can only have a maximum
of 1000 artifacts (some discussion captured here
https://github.com/orgs/community/discussions/165616).

This PR has two changes:

1. I have added a CI that will remove assets older than 60 days -- we
release 4 assets daily (3 for linux + 1 for windows). So technically
this can be increased to 250 days but I don't know if there is a need to
keep assets that old. I have chosen 60 days since I recall seeing
somewhere that's how long PyTorch keep their nightly releases but I
couldn't find that now.

An alternative to keeping all the assets will probably be to release the
assets with new name daily but seems overkill to me. Open to other
suggestions as well.

2. Added the `artifactErrorsFailBuild` flag on the release CIs which I
think will cause the builds to fail if the upload step fails --
currently the build doesn't fail which is why I think we didn't notice
this earlier.